refactor(ee): share credential_json helper across perm-sync connectors
Nine perm-sync modules inlined an identical
`cc_pair.credential.credential_json.get_value(apply_mask=False)` unwrap and
canvas kept it as a private helper. Promote it to a single shared
`ee/onyx/external_permissions/utils.py::credential_json` and migrate every
duplicator (canvas, github, gmail, google_drive, jira, salesforce, sharepoint,
slack, teams). Behavior-preserving.
