+++ backend/ee/onyx/external_permissions/canvas/doc_sync.py
@@ -1,8 +1,8 @@
 from collections.abc import Generator
 
-from ee.onyx.external_permissions.canvas.utils import credential_json
 from ee.onyx.external_permissions.perm_sync_types import FetchAllDocumentsFunction
 from ee.onyx.external_permissions.perm_sync_types import FetchAllDocumentsIdsFunction
+from ee.onyx.external_permissions.utils import credential_json
 from ee.onyx.external_permissions.utils import generic_doc_sync
 from onyx.access.models import ElementExternalAccess
 from onyx.configs.constants import DocumentSource
