+++ backend/ee/onyx/external_permissions/github/doc_sync.py
@@ -15,6 +15,7 @@ from ee.onyx.external_permissions.github.utils import get_repository_visibility
 from ee.onyx.external_permissions.github.utils import GitHubVisibility
 from ee.onyx.external_permissions.perm_sync_types import FetchAllDocumentsFunction
 from ee.onyx.external_permissions.perm_sync_types import FetchAllDocumentsIdsFunction
+from ee.onyx.external_permissions.utils import credential_json
 from onyx.access.models import DocExternalAccess
 from onyx.access.utils import build_ext_group_name_for_onyx
 from onyx.configs.constants import DocumentSource
@@ -50,12 +51,7 @@ def github_doc_sync(
         **cc_pair.connector.connector_specific_config
     )
 
-    credential_json = (
-        cc_pair.credential.credential_json.get_value(apply_mask=False)
-        if cc_pair.credential.credential_json
-        else {}
-    )
-    github_connector.load_credentials(credential_json)
+    github_connector.load_credentials(credential_json(cc_pair))
     logger.info("GitHub connector credentials loaded successfully")
 
     if not github_connector.github_client:
