+++ backend/ee/onyx/external_permissions/slack/doc_sync.py
@@ -7,6 +7,7 @@ from ee.onyx.external_permissions.perm_sync_types import FetchAllDocumentsIdsFun
 from ee.onyx.external_permissions.slack.channel_access import get_channel_access
 from ee.onyx.external_permissions.slack.utils import fetch_team_user_emails
 from ee.onyx.external_permissions.slack.utils import fetch_user_id_to_email_map
+from ee.onyx.external_permissions.utils import credential_json
 from onyx.access.models import DocExternalAccess
 from onyx.access.models import ExternalAccess
 from onyx.connectors.credentials_provider import OnyxDBCredentialsProvider
@@ -245,14 +246,10 @@ def slack_doc_sync(
     tenant_id = get_current_tenant_id()
     provider = OnyxDBCredentialsProvider(tenant_id, "slack", cc_pair.credential.id)
     r = get_redis_client(tenant_id=tenant_id)
-    credential_json = (
-        cc_pair.credential.credential_json.get_value(apply_mask=False)
-        if cc_pair.credential.credential_json
-        else {}
-    )
+    creds = credential_json(cc_pair)
     slack_client = SlackConnector.make_slack_web_client(
         provider.get_provider_key(),
-        credential_json["slack_bot_token"],
+        creds["slack_bot_token"],
         SlackConnector.MAX_RETRIES,
         r,
     )
